Update UI branch to be up to date with the rest of the engine code

+++ engine/source/resource/shader/ShaderResource.cpp
@@ -40,12 +40,12 @@ void engine::Shader::LoadResource(const char* fileName)
 	glShaderSource(m_shader, 1, &fileData.m_fileContent, &fileData.m_size);
 	glCompileShader(m_shader);
 
-	int result;
+	int32 result;
 	glGetShaderiv(m_shader, GL_COMPILE_STATUS, &result);
 
 	if (!result)
 	{
-		const int infoLogLength = 2500;
+		const int32 infoLogLength = 2500;
 		char infoLog[infoLogLength];
 
 		glGetShaderInfoLog(m_shader, infoLogLength, nullptr, infoLog);
@@ -58,7 +58,7 @@ void engine::Shader::LoadResource(const char* fileName)
 	fileData.Clear();
 }
 
-uint32_t engine::Shader::GetShader(void) const noexcept
+uint32 engine::Shader::GetShader(void) const noexcept
 {
 	return m_shader;
 }
@@ -70,14 +70,14 @@ engine::EShaderType engine::Shader::GetShaderType(void) const noexcept
 
 engine::EShaderType engine::Shader::ShaderType(const char* fileName) const
 {
-	size_t pos = std::string(fileName).rfind('.');
+	uint64 pos = std::string(fileName).rfind('.');
 
 	if (pos == std::string::npos)
 	{
 		std::printf("Failed to process shader. Shader file does not have an extension.\n");
 		return EShaderType::INVALID_SHADER;
 	}
-
+	
 	switch (fileName[pos + 1])
 	{
 	case 'v':
